Adding remove option on client register

+++ service-order/service-order-ejb/src/main/java/br/com/codeshare/service/ClientService.java
@@ -50,7 +50,9 @@ public class ClientService{
 		if(phonesToBeRemove != null){
 			for(Phone phone : phonesToBeRemove){
 				if(soService.findSoByPhoneId(phone.getId()).isEmpty()){
-					phoneService.remove(phone);
+					if(phone.getId()!=null){
+						phoneService.remove(phone);
+					}
 				}
 				else{
 					throw new BusinessException(ErrorCode.PHONE_HAS_SERVICE_ORDER.getErrorCode());
